pyproject.toml
[project]
name = "fastapi-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "fastapi>=0.129.0",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "ruff>=0.15.1",
    "ty>=0.0.17",
]
[project.scripts]
dev = "app.main:dev"
start = "app.main:start"

[tool.uv]
package = true
